GPUV-17 Merge con la rama "develop".
