+++ app/src/main/java/com/example/gestrenacer/repository/UserRepositorio.kt
@@ -25,29 +25,11 @@ import java.util.concurrent.TimeUnit
 import javax.inject.Inject
 
 class UserRepositorio @Inject constructor(
-    private val smsService: SmsService,
-    private val context: Context) {
+    private val smsService: SmsService) {
 
     private val auth: FirebaseAuth = FirebaseAuth.getInstance()
     private val db: FirebaseFirestore = FirebaseFirestore.getInstance()
     private val usersCollection = db.collection("users")
-    private val sharedPreferences: SharedPreferences = context.getSharedPreferences("user_prefs", Context.MODE_PRIVATE)
-
-
-    fun saveUserRole(role: String) {
-        sharedPreferences.edit().putString("user_role", role).apply()
-    }
-
-
-    fun getUserRole(): String? {
-        return sharedPreferences.getString("user_role", null)
-    }
-
-
-    fun clearUserRole() {
-        sharedPreferences.edit().remove("user_role").apply()
-    }
-
     suspend fun getUsers(
         filtroSexo: List<String>, filtroEstCivil: List<String>,
         filtroLlamado: List<String>, fechaInicial: Timestamp,
@@ -138,7 +120,6 @@ class UserRepositorio @Inject constructor(
                 val rol = document.getString("rol")
                     ?: "Feligrés"  // Rol predeterminado si no se encuentra
                 if (rol != "Feligrés") {
-                    saveUserRole(rol)
                     rol
                 } else {
                     null
