+++ app/src/main/java/com/example/gestrenacer/view/fragment/PendingFragment.kt
@@ -138,10 +138,9 @@ class PendingFragment : Fragment(), Recargable {
         binding.btnFiltrar.setOnClickListener{
             val listFiltros = userViewModel.filtros.value as List<List<String>>
             val listOrden = userViewModel.orden.value as List<String>
-            val role = userViewModel.rol.value as String
             val modalBottomSheet = ModalBottomSheet(
                 userViewModel::getFeligreses, listFiltros,
-                listOrden, groupViewModel, setAppliedFilters, role
+                listOrden, groupViewModel, setAppliedFilters
             )
             modalBottomSheet.show(requireActivity().supportFragmentManager, ModalBottomSheet.TAG)
         }
