+++ app/src/main/java/com/example/gestrenacer/view/fragment/SmsFragment.kt
@@ -1,26 +1,45 @@
 package com.example.gestrenacer
 
+import android.app.AlertDialog
 import android.os.Bundle
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.core.view.isVisible
 import androidx.core.widget.addTextChangedListener
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
 import androidx.navigation.fragment.findNavController
+import androidx.recyclerview.widget.DividerItemDecoration
+import androidx.recyclerview.widget.LinearLayoutManager
 import com.example.gestrenacer.databinding.FragmentSmsBinding
+import com.example.gestrenacer.models.Group
+import com.example.gestrenacer.models.Plantilla
 import com.example.gestrenacer.view.MainActivity
+import com.example.gestrenacer.view.adapter.GroupAdapter
+import com.example.gestrenacer.view.adapter.PlantillaAdapter
 import com.example.gestrenacer.view.modal.DialogUtils
+import com.example.gestrenacer.viewmodel.GroupViewModel
+import com.example.gestrenacer.viewmodel.PlantillaViewModel
 import com.example.gestrenacer.viewmodel.SmsViewModel
+import com.google.firebase.Timestamp
 import dagger.hilt.android.AndroidEntryPoint
+import java.text.SimpleDateFormat
+import java.util.Date
+import java.util.Locale
+import java.util.UUID
 
 @AndroidEntryPoint
 class SmsFragment : Fragment() {
 
     private lateinit var binding: FragmentSmsBinding
     private val smsViewModel: SmsViewModel by viewModels()
+    private val plantillaViewModel: PlantillaViewModel by viewModels()
+    private val groupViewModel: GroupViewModel by viewModels()
+    private lateinit var listaDePlantillas: MutableList<Plantilla>
+    private lateinit var plantillaAdapter: PlantillaAdapter
 
 
     override fun onCreateView(
@@ -35,7 +54,9 @@ class SmsFragment : Fragment() {
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
+        groupViewModel.getGroups()
         iniciarComponentes()
+        plantillaViewModel.obtenerPlantillas()
     }
 
     private fun iniciarComponentes() {
@@ -47,13 +68,16 @@ class SmsFragment : Fragment() {
         iniciarFiltros()
         iniciarCantSms()
         iniciarUsuarios()
+        initGroupsAutocomplete()
+        initializeRecyclerView()
         observerProgress()
         observerOperacion()
         observerGuardado()
         observerGrupoActivado()
+        observerPlantillas()
     }
 
-    private fun iniciarToolbar(){
+    private fun iniciarToolbar() {
         val activity = requireActivity() as MainActivity
         activity.visibilidadBottomBar(false)
         binding.toolbar.lblToolbar.text = getString(R.string.lblMensajeria)
@@ -71,7 +95,7 @@ class SmsFragment : Fragment() {
     private fun iniciarUsuarios() {
         val lista = requireArguments().getStringArrayList("usuarios")?.toMutableList()
 
-        Log.d("iniciar usuario",lista.toString())
+        Log.d("iniciar usuario", lista.toString())
         smsViewModel.iniciarUsuarios(lista as MutableList<String>)
     }
 
@@ -132,6 +156,14 @@ class SmsFragment : Fragment() {
         }
     }
 
+    private fun observerPlantillas() {
+        plantillaViewModel.plantillas.observe(viewLifecycleOwner) { plantillas ->
+            plantillas?.let {
+                updatePlantillaList(it)
+            }
+        }
+    }
+
     private fun mostrarChip(filtro: String) {
         val aux = filtro.split(" ")
         when (aux[0]) {
@@ -156,7 +188,7 @@ class SmsFragment : Fragment() {
     }
 
     private fun manejadorTxtSms() {
-        val list = listOf(binding.txtSms,binding.txtPlantilla)
+        val list = listOf(binding.txtSms, binding.txtPlantilla)
 
         list.forEach { i ->
             i.addTextChangedListener {
@@ -165,16 +197,125 @@ class SmsFragment : Fragment() {
         }
     }
 
-    private fun validarTexto(){
+    private fun validarTexto() {
         val cant = smsViewModel.usuarios.value?.size as Int
         val grupos = smsViewModel.grupoActivado.value as Boolean
         val plantilla = smsViewModel.guardado.value as Boolean
         val nomPlantilla = binding.txtPlantilla.text.toString().isNotEmpty()
         val txtSms = binding.txtSms.text.toString().isNotEmpty()
-
         binding.btnEnviar.isEnabled = (txtSms && (cant > 0 || grupos) && ((plantilla && nomPlantilla) || !plantilla))
 
         cambiarColorBtnEnviar()
+    }
+
+    private fun initGroupsAutocomplete() {
+        val autoCompleteEditText = binding.groupsAutoCompleteTv
+
+        autoCompleteEditText.visibility = View.VISIBLE
+
+        groupViewModel.listaGroups.observe(viewLifecycleOwner) { groups ->
+            val showDetailsDialog: (Group) -> Unit = { group ->
+                val dialog = AlertDialog.Builder(requireContext())
+                dialog.setTitle(group.nombre)
+                var message = "Este grupo incluye las siguientes categorías:"
+                for (c in group.checkboxfilters) {
+                    message += "\n - $c"
+                }
+                message += "\n\nY acota las edades en los siguientes rangos:\n"
+
+                var fechaInicial = "(Sin especificar)"
+                var fechaFinal = "(Sin especificar)"
+
+                val auxFechaInicial = group.datesfilters[0].toDate()
+                val auxFechaFinal = group.datesfilters[1].toDate()
+
+                if ((auxFechaInicial.date != 1) and (auxFechaInicial.month != 0) and (auxFechaInicial.year != 0)) {
+                    fechaInicial = formatearFecha(group.datesfilters[0])
+                }
+
+                if ((auxFechaFinal.date != 0) and (auxFechaFinal.month != 12) and (auxFechaFinal.year != 300)) {
+                    fechaFinal = formatearFecha(group.datesfilters[1])
+                }
+
+                message += "${fechaInicial} hasta ${fechaFinal}"
+
+                dialog.setMessage(message)
+                dialog.setNegativeButton("Cancelar", null)
+                dialog.setPositiveButton("Borrar") { _, _ ->
+                    groupViewModel.deleteGroup(group)
+                    groupViewModel.getGroups()
+                    autoCompleteEditText.text.clear()
+                }
+                dialog.show()
+            }
+
+            val adapter = GroupAdapter(
+                context = requireContext(),
+                allGroups = groups,
+                autoCompleteTextView = autoCompleteEditText,
+                showDetailsDialog = showDetailsDialog
+            )
+
+            autoCompleteEditText.setAdapter(adapter)
+        }
+    }
+
+    private fun initializeRecyclerView() {
+        listaDePlantillas = mutableListOf()
+
+        plantillaAdapter = PlantillaAdapter(listaDePlantillas,
+            onPlantillaClick = { plantilla -> onPlantillaClick(plantilla) },
+            onPlantillaLongClick = { plantilla -> showPlantillaDetailsDialog(plantilla) }
+        )
+
+        binding.recyclerViewPlantillas.layoutManager = LinearLayoutManager(context)
+        binding.recyclerViewPlantillas.adapter = plantillaAdapter
+
+        val dividerItemDecoration = DividerItemDecoration(
+            binding.recyclerViewPlantillas.context,
+            LinearLayoutManager.VERTICAL
+        )
+        binding.recyclerViewPlantillas.addItemDecoration(dividerItemDecoration)
+    }
+
+    private fun onPlantillaClick(plantilla: Plantilla) {
+        binding.txtPlantilla.setText(plantilla.name)
+        binding.txtSms.setText(plantilla.message)
+    }
+
+    private fun updatePlantillaList(plantillas: List<Plantilla>) {
+        listaDePlantillas.clear()
+        listaDePlantillas.addAll(plantillas)
+        plantillaAdapter.notifyDataSetChanged()
+    }
+
+    private fun showPlantillaDetailsDialog(plantilla: Plantilla) {
+        val dialog = AlertDialog.Builder(requireContext())
+        dialog.setTitle(plantilla.name)
+        dialog.setMessage(plantilla.message)
+        dialog.setNegativeButton("Cancelar", null)
+        dialog.setPositiveButton("Borrar") { _, _ ->
+            // Crear una lista con la plantilla a eliminar
+            val plantillasAEliminar = listOf(plantilla)
+
+            // Lógica para borrar la plantilla
+            plantillaViewModel.eliminarPlantillas(plantillasAEliminar) // Llama al método de eliminar en el ViewModel
+        }
+        dialog.show()
+    }
+
+    private fun crearPlantilla() {
+        val mensaje = binding.txtSms.text.toString().trim()
+        val nombrePlantilla = binding.txtPlantilla.text.toString().trim()
+        // Crear objeto Plantilla con un ID único
+        val nuevaPlantilla = Plantilla(
+            id = UUID.randomUUID().toString(), // Genera un ID único
+            name = nombrePlantilla,
+            message = mensaje
+        )
+        plantillaViewModel.crearPlantilla(nuevaPlantilla)
+        binding.txtSms.text?.clear()
+        binding.txtPlantilla.text?.clear()
     }
 
     private fun manejadorBtnVolver() {
@@ -187,12 +328,16 @@ class SmsFragment : Fragment() {
         if (binding.btnEnviar.isEnabled) {
             binding.btnEnviar.setBackgroundColor(resources.getColor(R.color.onSelectedColorBotBar))
         } else {
-            binding.btnEnviar.setTextColor(resources.getColor(R.color.btnDesactivado))
+            binding.btnEnviar.setBackgroundColor(resources.getColor(R.color.btnDesactivado))
         }
     }
 
     private fun manejadorBtnEnviar() {
         binding.btnEnviar.setOnClickListener {
+            val checked = binding.switchGuardSms.isChecked
+            if (checked) {
+                crearPlantilla()
+            }
             smsViewModel.enviarSms(binding.txtSms.text.toString())
         }
     }
@@ -202,5 +347,11 @@ class SmsFragment : Fragment() {
             binding.lblPlantilla.isEnabled = it
             smsViewModel.cambiarGuardado(it)
         }
+    }
+
+    private fun formatearFecha(timestamp: Timestamp): String {
+        val date: Date = timestamp.toDate()
+        val format = SimpleDateFormat("dd/MM/yyyy", Locale.getDefault())
+        return format.format(date)
     }
 }
